Move all flattening code into DisplayListFlattener
This change avoids as much refactoring as possible, but will allow
a great deal of code simplification in the future. With this change,
the lifetime of FrameBuilder and DisplayListFlattener no longer
overlap and every DisplayListFlattener consumes one old FrameBuilder
and produces a new one.

This is part of #2419.
